server/rag_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
chromadb==0.4.18
openai==1.3.7
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2

# Multimodal (CLIP) dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
Pillow>=10.0.0
ftfy>=6.1.0
